64 bits port
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@481 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/PlanarityTest.cpp
@@ -57,10 +57,10 @@ list<edge> PlanarityTest::getObstructionsEdges(SuperGraph *graph) {
 //=================================================================
 bool PlanarityTest::compute(SuperGraph *graph) { 
 
-  if (resultsBuffer.find((unsigned int)graph)!=resultsBuffer.end()) 
-    return resultsBuffer[(unsigned int)graph];
+  if (resultsBuffer.find((unsigned long)graph)!=resultsBuffer.end()) 
+    return resultsBuffer[(unsigned long)graph];
   else if(graph->numberOfNodes()==0){
-    resultsBuffer[(unsigned int)graph] = true;
+    resultsBuffer[(unsigned long)graph] = true;
     return true;
   }
   vector<edge> addedEdges;
